feat: add FeaturedCategorie component to homePage
