+++ src/apps/front-office/home/pages/HomePage/sections/FeaturedCategories/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./FeaturedCategories";
